Fix incorrect audio implementation when combining channels
This was the bug alluded to in the previous commit. Here the three per-voice samples were summed as unsigned bytes before being clamped to a signed range. Since the WSG3 produces signed samples stored in bytes (two's
complement), treating them as unsigned results in incorrect mixing. e.g. a sample of -80 stored as byte 176 was summed as +176.

Fixed by casting each channel sample to sbyte before adding.

+++ emulator/Platform/Platform.cs
@@ -337,9 +337,11 @@ namespace JustinCredible.PacEmu
 
             for (int i = 0; i < sourceSamples.Length; i++)
             {
-                var sampleChannel1 = samplesByChannel[i][0];
-                var sampleChannel2 = samplesByChannel[i][1];
-                var sampleChannel3 = samplesByChannel[i][2];
+                // Samples are signed values stored in bytes (two's complement); cast to
+                // sbyte so they are summed correctly rather than treated as unsigned.
+                var sampleChannel1 = (sbyte)samplesByChannel[i][0];
+                var sampleChannel2 = (sbyte)samplesByChannel[i][1];
+                var sampleChannel3 = (sbyte)samplesByChannel[i][2];
 
                 // Merge all three voices into one.
                 var sample = sampleChannel1 + sampleChannel2 + sampleChannel3;
